Update response handling for proxyReqRes call
