+++ lib/jsonwp-proxy/proxy.js
@@ -2,11 +2,13 @@ import _ from 'lodash';
 import { logger, util } from 'appium-support';
 import request from 'request-promise';
 import { getSummaryByCode } from '../jsonwp-status/status';
-import { errors, isErrorType, errorFromMJSONWPStatusCode, errorFromW3CJsonCode } from '../protocol/errors';
+import {
+  errors, isErrorType, errorFromMJSONWPStatusCode, errorFromW3CJsonCode
+} from '../protocol/errors';
 import BaseDriver from '../basedriver/driver';
 import { routeToCommandName } from '../protocol/routes';
 import ProtocolConverter from './protocol-converter';
-
+import { formatResponseValue } from '../protocol/helpers';
 
 const log = logger.getLogger('WD Proxy');
 // TODO: Make this value configurable as a server side capability
@@ -219,10 +221,10 @@ class JWProxy {
   }
 
   getProtocolFromResBody (resObj) {
-    if (util.hasValue(resObj.status)) {
+    if (_.isInteger(resObj.status)) {
       return MJSONWP;
     }
-    if (util.hasValue(resObj.value)) {
+    if (!_.isUndefined(resObj.value)) {
       return W3C;
     }
   }
@@ -306,21 +308,18 @@ class JWProxy {
     // if the proxied response contains a sessionId that the downstream
     // driver has generated, we don't want to return that to the client.
     // Instead, return the id from the request or from current session
-    let isResBodyModified = false;
     if (_.has(resBodyObj, 'sessionId')) {
       const reqSessionId = this.getSessionIdFromUrl(req.originalUrl);
       if (reqSessionId) {
         log.info(`Replacing sessionId ${resBodyObj.sessionId} with ${reqSessionId}`);
         resBodyObj.sessionId = reqSessionId;
-        isResBodyModified = true;
       } else if (this.sessionId) {
         log.info(`Replacing sessionId ${resBodyObj.sessionId} with ${this.sessionId}`);
         resBodyObj.sessionId = this.sessionId;
-        isResBodyModified = true;
       }
     }
-    res.status(response.statusCode)
-      .send(isResBodyModified ? JSON.stringify(resBodyObj) : response.body);
+    resBodyObj.value = formatResponseValue(resBodyObj.value);
+    res.status(response.statusCode).send(JSON.stringify(resBodyObj));
   }
 }
