+++ lib/protocol/protocol.js
@@ -1,25 +1,21 @@
 import _ from 'lodash';
-import { logger, util } from 'appium-support';
+import { util } from 'appium-support';
 import { validators } from './validators';
-import { errors, isErrorType, errorFromMJSONWPStatusCode, errorFromW3CJsonCode,
-         getResponseForW3CError, getResponseForJsonwpError } from './errors';
+import {
+  errors, isErrorType, getResponseForW3CError, getResponseForJsonwpError,
+  errorFromMJSONWPStatusCode, errorFromW3CJsonCode,
+} from './errors';
 import { METHOD_MAP, NO_SESSION_ID_COMMANDS } from './routes';
-import { duplicateKeys } from '../basedriver/helpers';
 import B from 'bluebird';
 import BaseDriver from '../basedriver/driver';
-import LRU from 'lru-cache';
+import {
+  MJSONWP_ELEMENT_KEY, W3C_ELEMENT_KEY, formatResponseValue, formatStatus,
+} from './helpers';
+import SESSIONS_CACHE from './sessions-cache';
 
-const GENERIC_PROTOCOL = 'GENERIC';
-const mjsonwpLog = logger.getLogger('MJSONWP');
-const w3cLog = logger.getLogger('W3C');
-const genericProtocolLog = logger.getLogger(GENERIC_PROTOCOL);
-
-const JSONWP_SUCCESS_STATUS_CODE = 0;
 // TODO: Make this value configurable as a server side capability
 const LOG_OBJ_LENGTH = 1024; // MAX LENGTH Logged to file / console
 
-const MJSONWP_ELEMENT_KEY = 'ELEMENT';
-const W3C_ELEMENT_KEY = 'element-6066-11e4-a52e-4f735466cecf';
 const IMAGE_ELEMENT_PREFIX = 'appium-image-element-';
 
 const CREATE_SESSION_COMMAND = 'createSession';
@@ -35,69 +31,6 @@ const IMG_EL_URL_RE = new RegExp(
 );
 
 class Protocol {}
-
-class SessionsCache {
-  constructor (max) {
-    this._cache = new LRU({ max });
-  }
-
-  getLogger (sessionId, protocol) {
-    if (sessionId) {
-      if (this._cache.has(sessionId)) {
-        const value = this._cache.get(sessionId);
-        if (value.logger) {
-          return value.logger;
-        }
-        protocol = protocol || value.protocol;
-      }
-      // Always create a new logger instance for ids
-      // that are not in the current sessions list,
-      // so we can still see such ids as prefixes
-      return logger.getLogger(`${protocol || GENERIC_PROTOCOL} ` +
-        `(${sessionId.substring(0, Math.min(sessionId.length, 8))})`);
-    }
-
-    // Fall back to protocol name-only logger if session id is unknown
-    switch (protocol) {
-      case BaseDriver.DRIVER_PROTOCOL.W3C:
-        return w3cLog;
-      case BaseDriver.DRIVER_PROTOCOL.MJSONWP:
-        return mjsonwpLog;
-      default:
-        return genericProtocolLog;
-    }
-  }
-
-  getProtocol (sessionId) {
-    return (this._cache.get(sessionId) || {}).protocol;
-  }
-
-  putSession (sessionId, value) {
-    if (sessionId && value) {
-      this._cache.set(sessionId, {
-        protocol: value,
-        // We don't want to cache the logger instance for each random session id in the cache
-        // in order to save memory. Instead we only cache loggers for valid ids that
-        // are returned by `createSession` call and reset them after `deleteSession` is called
-        logger: this.getLogger(sessionId, value),
-      });
-    }
-    return value;
-  }
-
-  resetLogger (sessionId) {
-    if (this._cache.has(sessionId)) {
-      this._cache.get(sessionId).logger = null;
-    }
-  }
-}
-
-// This cache is useful when a session gets terminated
-// and removed from the sessions list in the umbrella driver,
-// but the client still tries to send a command to this session id.
-// So we know how to properly wrap the error message for it
-const SESSIONS_CACHE = new SessionsCache(100);
-
 
 function extractProtocol (driver, sessionId = null) {
   const dstDriver = _.isFunction(driver.driverForSession)
@@ -382,16 +315,7 @@ function buildHandler (app, method, path, spec, driver, isSessCmd) {
         }
       }
 
-      // If the MJSONWP element key format (ELEMENT) was provided, add a duplicate key (element-6066-11e4-a52e-4f735466cecf)
-      // If the W3C element key format (element-6066-11e4-a52e-4f735466cecf) was provided, add a duplicate key (ELEMENT)
-      if (driverRes) {
-        driverRes = duplicateKeys(driverRes, MJSONWP_ELEMENT_KEY, W3C_ELEMENT_KEY);
-      }
-
-      // convert undefined to null, but leave all other values the same
-      if (_.isUndefined(driverRes)) {
-        driverRes = null;
-      }
+      driverRes = formatResponseValue(driverRes);
 
       // delete should not return anything even if successful
       if (spec.command === DELETE_SESSION_COMMAND) {
@@ -410,10 +334,7 @@ function buildHandler (app, method, path, spec, driver, isSessCmd) {
         }
       }
 
-      // Response status should be the status set by the driver response.
-      if (currentProtocol !== BaseDriver.DRIVER_PROTOCOL.W3C) {
-        httpResBody.status = (_.isNil(driverRes) || _.isUndefined(driverRes.status)) ? JSONWP_SUCCESS_STATUS_CODE : driverRes.status;
-      }
+      httpResBody = formatStatus(httpResBody, currentProtocol);
       httpResBody.value = driverRes;
       SESSIONS_CACHE.getLogger(req.params.sessionId || newSessionId, currentProtocol).debug(`Responding ` +
         `to client with driver.${spec.command}() result: ${_.truncate(JSON.stringify(driverRes), {length: LOG_OBJ_LENGTH})}`);
